+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Look for credential secrets in all namespaces.
+
 ## [2.10.0] - 2022-03-21
 
 ### Added