+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.11.0] - 2022-06-14
+
 ### Changed
 
 - Look for credential secrets in all namespaces.
@@ -158,7 +160,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/giantswarm/azure-collector/compare/v2.10.0...HEAD
+[Unreleased]: https://github.com/giantswarm/azure-collector/compare/v2.11.0...HEAD
+[2.11.0]: https://github.com/giantswarm/azure-collector/compare/v2.10.0...v2.11.0
 [2.10.0]: https://github.com/giantswarm/azure-collector/compare/v2.9.0...v2.10.0
 [2.9.0]: https://github.com/giantswarm/azure-collector/compare/v2.8.0...v2.9.0
 [2.8.0]: https://github.com/giantswarm/azure-collector/compare/v2.7.1...v2.8.0